Some mistakes
I was going back to see if it was changed after I edited it the first time then realized that there was a chord I forgot to throw in and also a minor form mistake. Also, I threw in the lick that Trey plays at SPAC before the chorus as well because I feel like it is a super sweet transition between the two.